Components: Extract a reusable PostAuthors Component
