+++ editor/post-author/test/index.js
@@ -62,43 +62,6 @@ describe( 'PostAuthor', () => {
 	} );
 
 	describe( '#render()', () => {
-		it( 'should not render anything if the user doesn\'t have the right capabilities', () => {
-			let wrapper = shallow( <PostAuthor users={ users } user={ {} } /> );
-			expect( wrapper.type() ).toBe( null );
-			wrapper = shallow( <PostAuthor users={ users } user={
-				{ data: { capabilities: { publish_posts: false } } }
-			} /> );
-			expect( wrapper.type() ).toBe( null );
-		} );
-
-		it( 'should not render anything if users unknown', () => {
-			const wrapper = shallow( <PostAuthor users={ {} } user={ user } /> );
-
-			expect( wrapper.type() ).toBe( null );
-		} );
-
-		it( 'should not render anything if single user', () => {
-			const wrapper = shallow(
-				<PostAuthor users={ { data: users.data.slice( 0, 1 ) } } user={ user } />
-			);
-
-			expect( wrapper.type() ).toBe( null );
-		} );
-
-		it( 'should not render anything if single filtered user', () => {
-			const wrapper = shallow(
-				<PostAuthor users={ { data: users.data.slice( 0, 2 ) } } user={ user } />
-			);
-
-			expect( wrapper.type() ).toBe( null );
-		} );
-
-		it( 'should render select control', () => {
-			const wrapper = shallow( <PostAuthor users={ users } user={ user } /> );
-
-			expect( wrapper.find( 'select' ).length ).not.toBe( 0 );
-		} );
-
 		it( 'should update author', () => {
 			const onUpdateAuthor = jest.fn();
 			const wrapper = shallow(
